third_party/WebKit/LayoutTests/fast/canvas/canvas-context-attributes-default-value-expected.txt

Testing default value:
PASS actualContextAttributes.alpha is true

Testing undefined value:
PASS actualContextAttributes.alpha is true

Testing null value:
PASS actualContextAttributes.alpha is false

PASS successfullyParsed is true

TEST COMPLETE
